hardware: Replace current sense with a schottky, scrap blocking FET Q2
We discovered that we require a diode anyways, so to be a bit more
efficient, replace the fairly large current sense resistor with a
schottky. This required that we updated the voltage divider values to
reflect the new required voltage drop.

Future work here is to consider the worst case highest temperature we
want to support, and then add in a digipot to allow the micro to adjust
the voltage (and therefore the current) up from that point.